Interface to operate with different kinds of floats

+++ mandelbrot_renderer/wasm/golang/objects/size.go
@@ -1,6 +1,6 @@
 package objects
 
 type Size struct {
-	Width  int
-	Height int
+	Width  int64
+	Height int64
 }
